Refactor usb2can interface into own package
