+++ can/interfaces/usb2can/__init__.py
@@ -1,0 +1,1 @@
+from can.interfaces.usb2can.usb2canInterface import Usb2canBus
